Modularize Customer domain
