+++ src/main/api/customer/response/CreateCustomerResponseDTO.ts
@@ -1,0 +1,3 @@
+export interface CreateCustomerResponseDTO {
+    "user_full_name": string;
+}
